Struct smtp::client::Client

pub struct Client<S: Write + Read = SmtpStream> {
    // some fields omitted
}

Structure that implements the SMTP client

Methods

impl<S: Write + Read = SmtpStream> Client<S>

fn new<A: ToSocketAddrs>(addr: A) -> Client<S>

Creates a new SMTP client

It does not connects to the server, but only creates the Client

impl<S: Connector + Write + Read = SmtpStream> Client<S>

fn close(&mut self)

Closes the SMTP transaction if possible

fn connect(&mut self) -> SmtpResult

Connects to the configured server

fn is_connected(&mut self) -> bool

Checks if the server is connected using the NOOP SMTP command

fn command(&mut self, command: &str) -> SmtpResult

Sends an SMTP command

fn helo(&mut self, hostname: &str) -> SmtpResult

Send a HELO command and fills server_info

fn ehlo(&mut self, hostname: &str) -> SmtpResult

Sends a EHLO command and fills server_info

fn mail(&mut self, address: &str, options: Option<&str>) -> SmtpResult

Sends a MAIL command

fn rcpt(&mut self, address: &str) -> SmtpResult

Sends a RCPT command

fn data(&mut self) -> SmtpResult

Sends a DATA command

fn quit(&mut self) -> SmtpResult

Sends a QUIT command

fn noop(&mut self) -> SmtpResult

Sends a NOOP command

fn help(&mut self, argument: Option<&str>) -> SmtpResult

Sends a HELP command

fn vrfy(&mut self, address: &str) -> SmtpResult

Sends a VRFY command

fn expn(&mut self, address: &str) -> SmtpResult

Sends a EXPN command

fn rset(&mut self) -> SmtpResult

Sends a RSET command

fn auth_plain(&mut self, username: &str, password: &str) -> SmtpResult

Sends an AUTH command with PLAIN mecanism

fn auth_cram_md5(&mut self, username: &str, password: &str) -> SmtpResult

Sends an AUTH command with CRAM-MD5 mecanism

fn message(&mut self, message_content: &str) -> SmtpResult

Sends the message content