Struct smtp::sender::SenderBuilder

pub struct SenderBuilder {
    // some fields omitted
}

Contains client configuration

Methods

impl SenderBuilder

Builder for the SMTP Sender

fn new<A: ToSocketAddrs>(addr: A) -> SenderBuilder

Creates a new local SMTP client

fn localhost() -> SenderBuilder

Creates a new local SMTP client to port 25

fn hello_name(self, name: &str) -> SenderBuilder

Set the name used during HELO or EHLO

fn enable_connection_reuse(self, enable: bool) -> SenderBuilder

Enable connection reuse

fn connection_reuse_count_limit(self, limit: u16) -> SenderBuilder

Set the maximum number of emails sent using one connection

fn credentials(self, username: &str, password: &str) -> SenderBuilder

Set the client credentials

fn build(self) -> Sender

Build the SMTP client

It does not connects to the server, but only creates the Sender