Module smtp::toolsExperimental

Tools for common string manipulations

Functions

escape_crlf
Returns the string replacing all the CRLF with "<CRLF>"

escape_dot
Returns the string after adding a dot at the beginning of each line starting with a dot

get_first_word
Returns the first word of a string, or the string if it contains no space

remove_trailing_crlf
Removes the trailing line return at the end of a string