Enum smtp::transaction::TransactionStateUnstable

pub enum TransactionState {
    Unconnected,
    Connected,
    HelloSent,
    MailSent,
    RecipientSent,
    DataSent,
}

Contains the state of the current transaction

Variants

Unconnected

No connection was established

Connected

The connection was successful and the banner was received

HelloSent

An HELO or EHLO was successful

MailSent

A MAIL command was successful send

RecipientSent

At least one RCPT command was sucessful

DataSent

A DATA command was successful

Methods

impl TransactionState

fn new() -> TransactionState

Returns the initial state

fn is_allowed(&self, command: &Command) -> bool

Tests if the given command is allowed in the current state

fn next_state(&self, command: &Command) -> Option<TransactionState>

Returns the state resulting of the given command

A None return value means the comand is not allowed.