Struct smtp::client::ClientExperimental

pub struct Client<S = TcpStream> {
    // some fields omitted
}

Structure that implements the SMTP client

Methods

impl<S = TcpStream> Client<S>

fn new<A: ToSocketAddr>(addr: A) -> Client<S>

Creates a new SMTP client

It does not connects to the server, but only create the Client

fn localhost() -> Client<S>

Creates a new local SMTP client to port 25

It does not connects to the server, but only create the Client

fn set_hello_name(&mut self, name: &str)

Set the name used during HELO or EHLO

fn set_enable_connection_reuse(&mut self, enable: bool)

Set the maximum number of emails sent using one connection

fn set_connection_reuse_count_limit(&mut self, count: uint)

Set the maximum number of emails sent using one connection

fn set_configuration(&mut self, configuration: Configuration)

Set the client configuration

impl<S: Connecter + ClientStream + Clone = TcpStream> Client<S>

fn close(&mut self)

Closes the SMTP transaction if possible

fn reset(&mut self)

Reset the client state

fn send<T: SendableEmail>(&mut self, email: T) -> SmtpResult

Sends an email

fn connect(&mut self) -> SmtpResult

Connects to the configured server

fn is_connected(&mut self) -> bool

Checks if the server is connected using the NOOP SMTP command

fn helo(&mut self) -> SmtpResult

Send a HELO command and fills server_info

fn ehlo(&mut self) -> SmtpResult

Sends a EHLO command and fills server_info

fn mail(&mut self, from_address: &str) -> SmtpResult

Sends a MAIL command

fn rcpt(&mut self, to_address: &str) -> SmtpResult

Sends a RCPT command

fn data(&mut self) -> SmtpResult

Sends a DATA command

fn message(&mut self, message_content: &str) -> SmtpResult

Sends the message content

fn quit(&mut self) -> SmtpResult

Sends a QUIT command

fn rset(&mut self) -> SmtpResult

Sends a RSET command

fn noop(&mut self) -> SmtpResult

Sends a NOOP command

fn vrfy(&mut self, to_address: &str) -> SmtpResult

Sends a VRFY command

fn expn(&mut self, list: &str) -> SmtpResult

Sends a EXPN command